restored foundry priority check
The line that was removed here ensured that if you were pasting something you copied *in Foundry* it didn't also paste something you copied from *outside Foundry*.

The module broke because `_copy` is not a property that exists anymore- but there is an alternative. This change restores the original behavior via the `clipboard` property.